+++ src/main/cxx/optimize.h
@@ -551,8 +551,7 @@ namespace especia {
 
             // Check if the optimization is completed
             for (size_t i = 0, ii = 0; i < n; ++i, ii += n + 1) {
-                optimized = (sqr(step_size) * C[ii] <
-                             sqr(accuracy_goal * xw[i]) + 1.0 / max_covariance_matrix_condition);
+                optimized = (sqr(step_size) * C[ii] < sqr(accuracy_goal * xw[i]) + 1.0 / max_covariance_matrix_condition);
                 if (!optimized)
                     break;
             }